move functions related to get_valid_jump_destinations
