+++ cairo/tests/legacy/utils/test_utils.cairo
@@ -91,56 +91,6 @@ func test__bytes_used_128{range_check_ptr}(output_ptr: felt*) {
     return ();
 }
 
-func test__finalize_jumpdests{range_check_ptr}() {
-    alloc_locals;
-
-    local bytecode: felt*;
-    local valid_jumpdests_start: DictAccess*;
-    local valid_jumpdests: DictAccess*;
-    %{
-        from starkware.cairo.common.dict import DictTracker
-        from tests.utils.helpers import flatten
-        from ethereum.cancun.vm.runtime import get_valid_jump_destinations
-
-        memory[fp] = segments.add()
-        segments.write_arg(memory[fp], program_input["bytecode"])
-
-        data = {k: 1 for k in get_valid_jump_destinations(program_input["bytecode"])}
-
-        base = segments.add()
-        segments.load_data(
-            base,
-            flatten([[int(k), 1, 1] for k in data.keys()])
-        )
-        __dict_manager.trackers[base.segment_index] = DictTracker(
-            data=data,
-            current_ptr=(base + len(data) * 3),
-        )
-        memory[fp + 1] = base
-        memory[fp + 2] = base + len(data) * 3
-    %}
-
-    let (sorted_keys_start, sorted_keys_end) = dict_squash(valid_jumpdests_start, valid_jumpdests);
-
-    Helpers.finalize_jumpdests(0, sorted_keys_start, sorted_keys_end, bytecode);
-
-    return ();
-}
-
-func test__assert_valid_jumpdest{range_check_ptr}() {
-    alloc_locals;
-    tempvar bytecode: felt*;
-    tempvar valid_jumpdest: DictAccess*;
-    %{
-        ids.bytecode = segments.add()
-        segments.write_arg(ids.bytecode, program_input["bytecode"])
-        ids.valid_jumpdest = segments.add()
-        segments.write_arg(ids.valid_jumpdest.address_, program_input["valid_jumpdest"])
-    %}
-    Helpers.assert_valid_jumpdest(0, bytecode, valid_jumpdest);
-    return ();
-}
-
 func test__split_word{range_check_ptr}() -> felt* {
     alloc_locals;
     local value: felt;
