+++ cairo/tests/legacy/utils/test_utils.py
@@ -1,13 +1,10 @@
 from math import ceil
 
 import pytest
-from ethereum.cancun.vm.runtime import get_valid_jump_destinations
-from ethereum_types.bytes import Bytes
-from hypothesis import example, given, settings
+from hypothesis import given, settings
 from hypothesis import strategies as st
 
 from cairo_addons.testing.errors import cairo_error
-from tests.utils.solidity import get_contract
 
 pytestmark = pytest.mark.python_vm
 
@@ -79,47 +76,6 @@ def test_bytes_to_uint256(cairo_run, word):
 def test_should_return_bytes_used_in_128_word(cairo_run, word):
     bytes_length = (word.bit_length() + 7) // 8
     assert bytes_length == cairo_run("test__bytes_used_128", word=word)
-
-
-class TestFinalizeJumpdests:
-    @given(bytecode=...)
-    @example(bytecode=get_contract("Counter", "Counter").bytecode_runtime)
-    def test_should_pass(self, cairo_run, bytecode: Bytes):
-        cairo_run(
-            "test__finalize_jumpdests",
-            bytecode=list(bytecode),
-            valid_jumpdests=get_valid_jump_destinations(bytecode),
-        )
-
-
-class TestAssertValidJumpdest:
-    @pytest.mark.parametrize(
-        "jumpdest",
-        get_valid_jump_destinations(
-            get_contract("Counter", "Counter").bytecode_runtime
-        ),
-    )
-    def test_should_pass_on_valid_jumpdest(self, cairo_run, jumpdest):
-        cairo_run(
-            "test__assert_valid_jumpdest",
-            bytecode=list(get_contract("Counter", "Counter").bytecode_runtime),
-            valid_jumpdest=[int(jumpdest), 1, 1],
-        )
-
-    def test_should_raise_if_jumpdest_but_false(self, cairo_run):
-        with cairo_error("assert_valid_jumpdest: invalid jumpdest"):
-            cairo_run(
-                "test__assert_valid_jumpdest", bytecode=[0x5B], valid_jumpdest=[0, 0, 0]
-            )
-
-    @pytest.mark.parametrize("push", list(range(0x60, 0x80)))
-    def test_should_raise_if_jumpdest_is_push_arg(self, cairo_run, push):
-        with cairo_error("assert_valid_jumpdest: invalid jumpdest"):
-            cairo_run(
-                "test__assert_valid_jumpdest",
-                bytecode=[push] + (push - 0x5F) * [0x5B],
-                valid_jumpdest=[push - 0x5F, 1, 1],
-            )
 
 
 class TestSplitWord:
